fix unionfix fields on metadata and browse
